Rename 'Command' to 'CommandBase' to keep from clashing with the 'Command' DB Model
